Rewrite Worker for per-task fibers with free + ready lists
Replaces the fixed pool of 4 long-lived FiberSlots with grow-on-demand
per-task Fibers backed by a free list (recycled fibers waiting for new
tasks) and a ready list (suspended fibers whose blocker just resolved).
Each task gets a fresh Fiber.reset; each `.busy` yield's wake_fn pushes
the fiber onto the ready list and nudges its worker.

Widens Thunk.claimer to u32 = (u8 worker_id, u24 fiber_id) so we have
~16M fiber-id headroom per worker instead of 256.

Drops MAIN_THREAD_SLOT — there is no longer a "main thread without a
fiber" path; every code path runs inside a fiber, and the default
fiber id used by initVm is overwritten by Worker when binding a VM to
a fiber.

NixOS minimal wall-clock is ~3.8s vs ~3.3s on the slot model — the
ready-list SpinMutex contention and per-task allocation cost a small
amount on DAG-bound workloads. The architectural payoff is unbounded
in-flight fibers per worker (limited by memory, not a fixed N), which
will pay off on wide strict-rendering workloads where the slot model
would starve. Test-suite RSS drops from 317M to 43M since no pool is
pre-allocated.

+++ src/eval.zig
@@ -439,7 +439,7 @@ pub const Evaluator = struct {
         // main thread can yield on a `.busy` thunk; nested invocations
         // (imports, scoped imports) run synchronously on the existing
         // fiber's stack — they share the outer VM's claim identity via
-        // the default `MAIN_THREAD_SLOT` claimer that `initVm` hands out.
+        // default placeholder claimer that `initVm` hands out.
         if (scope == null and source_path == null) {
             self.progressBegin(.evaluate, subject);
             defer self.progressEnd(.evaluate, subject);
@@ -852,7 +852,7 @@ fn helperLoop(helper_idx: u8, sched: *Scheduler, ev: *Evaluator) void {
     worker.run();
 }
 
-fn initVmForWorkerSlot(ctx: *anyopaque, worker_id: u8, _: u8) anyerror!VM {
+fn initVmForWorkerSlot(ctx: *anyopaque, worker_id: u8, _: u32) anyerror!VM {
     const ev: *Evaluator = @ptrCast(@alignCast(ctx));
     return ev.initVm(worker_id);
 }
